Change data types regarding sensor readings and sensor/actuator pins from signed to unsigned

+++ firmware/atmega/source/moisture.c
@@ -21,22 +21,21 @@ void triggerMoistureRead(void) {
     readSensors = true;
 }
 
-int16_t moistureAverage(int8_t sensorPin) {
-    int16_t sum = 0;
-    for (int i = 0; i < SENSOR_READINGS; i++) {
+uint16_t moistureAverage(uint8_t sensorPin) {
+    uint16_t sum = 0;
+    for (uint8_t i = 0; i < SENSOR_READINGS; i++) {
         sum += adcRead(sensorPin);
         //TODO: implement variance check (if over a certain variance discard measurements for this cycle)
-        //TODO: maybe add a small delay using watchdog here?
         //TODO IMPORTANT: if value < 1 or > 1022, discard that measurement
     }
     return sum / SENSOR_READINGS;
 }
 
-int16_t moistureRead(void) {
+uint16_t moistureRead(void) {
     uint16_t overallSum = 0;
 
-    for (int8_t i = 0; i < SENSORS_NUM; i++) {
-        int16_t sensorAverage = moistureAverage(sensorPins[i]);
+    for (uint8_t i = 0; i < SENSORS_NUM; i++) {
+        uint16_t sensorAverage = moistureAverage(sensorPins[i]);
         //TODO: if a sensor gives 0, discard that sensor entirely (means it is disconnected or not working)
         overallSum += sensorAverage;
     }
